Add Hive.Client.Coordinate
