+++ src/Hive/Client/Board.hs
@@ -7,19 +7,16 @@ module Hive.Client.Board
 
 import qualified Data.Map.Strict as Map
 import Data.Monoid ((<>))
-import Data.Text (Text)
 import qualified Data.Text as Text
 import Reflex.Dom.Core
 
+import Hive.Client.Coordinate (asPoints, toSVGCoord)
 import Hive.Client.Piece
 import Hive.Client.SVG (svgAttr)
 import Hive.Core.Board (Board, PlayerPiece, coordinateMap)
 import Hive.Core.Coordinate (Coordinate)
 import Hive.Core.Piece (PieceType(..), pieceToType)
 import Hive.Core.Player (Player(..))
-
--- | An SVG coordinate corresponding to pieces of height 500px.
-type SVGCoordinate = (Double, Double)
 
 renderBoard :: MonadWidget t m => Board -> m ()
 renderBoard board = mapM_ (uncurry renderPiece) $ Map.toList $ normalize coordMap
@@ -57,17 +54,6 @@ renderPiece coord ((player, piece):_) = svgAttr "g" attrs $ do
 
 {- Helpers -}
 
--- | Convert a Coordinate to an SVG coordinate.
-toSVGCoord :: Coordinate -> SVGCoordinate
-toSVGCoord (x, y) = (x' * dx + padding, y' * dyPerY + x' * dyPerX + padding)
-  where
-    x' = fromIntegral x
-    y' = fromIntegral y
-    dx = 250 * sqrt 3
-    dyPerX = 250
-    dyPerY = -500
-    padding = 50
-
 -- | An SVG hexagon.
 hexagon :: MonadWidget t m => m ()
 hexagon = svgAttr "polygon" attrs $ pure ()
@@ -85,7 +71,3 @@ hexagon = svgAttr "polygon" attrs $ pure ()
       , (750 / sqrt 3 , 500)
       , (250  / sqrt 3, 500)
       ]
-
--- | Render the given coordinate as Text.
-asPoints :: SVGCoordinate -> Text
-asPoints (x, y) = Text.pack $ show x ++ "," ++ show y
